post-user relationship

+++ 22_APIREST/controllers/feed.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const {validationResult} = require('express-validator');
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -46,23 +47,35 @@ exports.postPost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
     }
-    const imageUrl = req.file.path.replace("\\" ,"/");
 
+    const imageUrl = req.file.path.replace("\\" ,"/");
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title, 
         content: content, 
         imageUrl: imageUrl,
-        creator: {name: 'Landeracio'},
+        creator: req.userId
     });
     post.save()
     .then(result => {
-        console.log(result);
+        return User.findById(req.userId);
+        
+    })
+    .then(user => {
+        creator = user;
+        //Link post to the user creator
+        user.posts.push(post);
+        return user.save();
         //201 = succes creating a resource
+        
+    })
+    .then(result => {
         res.status(201).json({
             message: 'Post created succesfully!',
-            post: result
+            post: post,
+            creator: {_id: creator._id, name: creator.name}
         });
     })
     .catch(err => {
@@ -124,6 +137,14 @@ exports.updatePost = (req, res, next) => {
             error.status = 404;
             throw error;
         }
+
+        if(post.creator.toString() !== req.userId){
+            const error = new Error('Not authorized');
+            //403 = authorization issues
+            error.statusCode = 403;
+            throw error;
+        }
+
         if(imageUrl !== post.imageUrl){
             clearImage(post.imageUrl);
         }
@@ -154,12 +175,26 @@ exports.deletePost = (req, res, next) => {
             error.status = 404;
             throw error;
         }
+        if(post.creator.toString() !== req.userId){
+            const error = new Error('Not authorized');
+            //403 = authorization issues
+            error.statusCode = 403;
+            throw error;
+        }
         //Check logged in user
         clearImage(post.imageUrl);
         return Post.findByIdAndDelete(postId);
     })
     .then(result => {
-        console.log(result);
+        return User.findById(req.userId);
+        
+    })
+    .then(user => {
+        //Removes post deleted from user
+        user.posts.pull(postId);
+        return user.save();
+    })
+    .then(result => {
         res.status(200).json({message: 'Deleted post'});
     })
     .catch(err => {
